requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
djongo
django-cors-headers
Pillow
